This is the first spiderman tutorial
This tutorial was having problems due to steamvr being updated. There is code in the web script that uses the public class SteamVR_TrackedObject which can't be used anymore since steamvr has the new input system. This made me confused as I don't know how to fix the code and implement it to the controllers. I'll still try find more tutorials to help me.